refactor: 서비스 구조 단순화
